refactor(api): PR7 — scanner runtime extracted, btc_api.py becomes bootstrap
Move execute_scan_for_symbol, scanner_loop, start_scanner_thread,
check_pending_signal_outcomes (writing version), _scanner_state, and
_BACKUP_INTERVAL_CYCLES/_backup_cycles_since_last from btc_api.py to
scanner/runtime.py.

scanner/runtime.py uses lazy imports (inside function bodies) for
api.signals, api.positions, api.config to stay within the scanner/*
import boundary rules validated by test_import_boundaries.py. Only
api.telegram is imported at top level (allowed as a service layer).

Remove all temporary re-exports from btc_api.py. Trim to exactly 200 LOC:
FastAPI app, lifespan (init_db + start_scanner_thread), CORS middleware,
mount of 8 domain routers, and 5 bootstrap-level inline routes (/,
/symbols, /status, /scan, /webhook/test).

Re-exports kept in btc_api.py are strictly for test compatibility (tests
monkeypatch btc_api.DB_FILE, btc_api.scan, btc_api.notify, etc.) —
commented with "noqa: F401" and "tests patch/mutate" annotations.

db/connection.py: keep lazy btc_api import in _resolve_db_file (inside
function body, boundary-exempt) to honour 50+ monkeypatch.setattr(btc_api,
"DB_FILE", ...) fixtures without touching all test files. Update docstring.

tests/test_api.py: update TestExecuteScanForSymbol to patch
scanner.runtime.scan (execute_scan_for_symbol now lives there and uses its
own import, not btc_api's). Setup fixture patches api.signals /
api.positions directly since those are the modules lazily imported by
execute_scan_for_symbol at call time. Also add save_config to btc_api
re-exports (2 calibrator tests patch btc_api.save_config).

tests/test_import_boundaries.py: update comment about lazy btc_api import
in db/connection.py — retained post-PR7 to avoid touching 50+ fixtures.

Definition of Done (spec §8):
  - btc_api.py ≤ 200 LOC: 200 exactly
  - 10 modules in api/: config, deps, health, kill_switch, notifications,
    ohlcv, positions, signals, telegram, tune
  - 4 modules in db/: connection, positions, schema, signals
  - scanner/runtime.py exists with all 4 functions + state
  - strategy/constants.py is single source for indicator constants
  - 1029 tests pass (no regression)

btc_scanner.py post-#186 cleanup tracked as follow-up issue.
